requirements.txt
smbus-cffi
pyserial
datadog


soundmeter